requirements.txt
pydub
pdfplumber
send2trash
google-cloud-texttospeech
ffmpeg-python
